NodeNumber value can change type to int\long\double. NodeNumber ::GetInteger, ::GetLong and ::GetDouble work for all types of numbers. Fixed errors with float numbers. Removed some warnings in SuperCalc project.

+++ SuperCalc.cpp
@@ -244,7 +244,7 @@ char typeTest(int x, short y, char z, int d, long long u, float m, double k)
 	DWORD written;
 	char buf[64];
 	sprintf(buf, "%d %d %d %d %I64d %f %f", x, y, z, d, u, m, k);
-	WriteFile(conStdOut, buf, strlen(buf), &written, NULL); 
+	WriteFile(conStdOut, buf, (DWORD)strlen(buf), &written, NULL); 
 	return 12;
 }
